Move 'HTTPClientSpy' to its specific file
